+++ server/startup.jsx
@@ -1,8 +1,0 @@
-//This does pretty much FA
-//I don't know...
-
-console.log("Starting server...");
-
-Meteor.startup(function(){
-    //FM.resetTestData();
-})